blog paper style, dry nav bar
Summary:

Test Plan:
